Start work on action_cache_asset_store
