+++ pkg/storage/action_cache_asset_store.go
@@ -1,0 +1,15 @@
+package storage
+
+import "github.com/buildbarn/bb-storage/pkg/blobstore"
+
+type actionCacheAssetStore struct {
+	actionCache             blobstore.BlobAccess
+	maximumMessageSizeBytes int
+}
+
+func NewActionCacheAssetStore(actionCache blobstore.BlobAccess, maximumMessageSizeBytes int) AssetStore {
+	return &actionCacheAssetStore{
+		actionCache:             actionCache,
+		maximumMessageSizeBytes: maximumMessageSizeBytes,
+	}
+}
